requirements.txt
tk
Pillow
tqdm
